# test/librbd: use GTEST_SKIP macro to skip tests