@trociny 2023-04-27
This is the problem I mentioned. It seems to be fixed.
https://tracker.ceph.com/issues/43566